complete search service usecases and their tests

+++ backend/searchservice/internal/app/usecase/delete_ad_index_uc.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"errors"
 
-	pkgerrs "github.com/maket12/ads-service/backend/authservice/pkg/errs"
 	"github.com/maket12/ads-service/backend/searchservice/internal/app/dto"
 	ucerrs "github.com/maket12/ads-service/backend/searchservice/internal/app/errs"
 	"github.com/maket12/ads-service/backend/searchservice/internal/domain/port"
+	pkgerrs "github.com/maket12/ads-service/backend/searchservice/pkg/errs"
 )
 
 type DeleteAdIndexUC struct {
@@ -19,11 +19,16 @@ func NewDeleteAdIndexUC(adIndex port.AdIndexRepository) *DeleteAdIndexUC {
 }
 
 func (uc *DeleteAdIndexUC) Execute(ctx context.Context, in dto.DeleteAdIndexInput) error {
+	if in.ID == "" {
+		return ucerrs.ErrInvalidAdIndex
+	}
+
 	if err := uc.adIndex.Delete(ctx, in.ID); err != nil {
 		if errors.Is(err, pkgerrs.ErrObjectNotFound) {
 			return ucerrs.ErrAdIndexNotFound
 		}
 		return ucerrs.Wrap(ucerrs.ErrDeleteAdIndexES, err)
 	}
+
 	return nil
 }
